base: Support re-baking adn box-only baking
Optimised the `bake.sh` script to enable running it on an already baked
image to add images to it without adding new copies of existing images.

This is done by avoiding deletion of the Vagrant box files and instead
making hard-links between the copies of the Vagrant images in the
$VAGRANT_HOME directory to the copies in the libvirt storage pool.

Also enabled a 'box-only` bake mode that allows creating images that
contain cached cipies of Vagrant boxes without also containing
pre-configured VMs. This allows creating families of images that
contain different VM configurations but all share the layer that contain
the large Box image file.

Along the way also removed some packages from the base image that caused
security alerts to show up in quay.io.

Signed-off-by: Barak Korren <bkorren@redhat.com>