Add a `builder` image
Add files for creating a `builder` container image that can be used for
building images with Vagrant environments in them based on the `base`
image.

The builder image is essentially the `base` image with `skopeo` and
`buildah` added and with a `derive.sh` script that lets one `bake`
images without having to run the image being built as a container inside
the image that is building it.

The builder image image accomplishes this feat by bind-mounting the
locations inside the image being build that store VM and Vagrant-related
information onto itself and them running the `bake.sh` script ao tha
that baked image ends up being stored in the built image.

The builder image also supports pushing the built images to a remote
container repository.

During the build process, the builder image stores intermediate
container data in a hidden directory in the /workspace directory which
should be a mounted volume. Since it is likely that the Vagrant
application data will also come from that directory, the `bake.sh`
script in the base image had been changed to not copy hidden files into
the Vagrant environment.

Signed-off-by: Barak Korren <bkorren@redhat.com>