Add general build script - build.sh
Added a build script that is meant to build all the images in this
repository automatically.

The script is meant to be run inside a builder container, if it is not
run inside such a container the script attempts to launch such a
container and invoke itself inside it.

This script accepts parameters that tell it where to push the images it
builds. Those are similar to the parameters one acn pass to
`publish.sh`.

Since the build process is likely to generate files in the source
directory, added those files to the `.gitgnore` file.

Signed-off-by: Barak Korren <bkorren@redhat.com>